requirements.txt
fastapi
uvicorn[standard]
httpx
feedparser
google-genai
apscheduler
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
python-multipart